# [jf+exec] Jellyfish batch update to leverage execution result