--- a/storage/jellyfish-merkle/src/lib.rs
+++ b/storage/jellyfish-merkle/src/lib.rs
@@ -74,7 +74,6 @@ mod jellyfish_merkle_test;
 pub mod metrics;
 #[cfg(any(test, feature = "fuzzing"))]
 mod mock_tree_store;
-mod nibble_path;
 pub mod node_type;
 pub mod restore;
 #[cfg(any(test, feature = "fuzzing"))]
@@ -83,12 +82,14 @@ mod tree_cache;
 
 use anyhow::{bail, ensure, format_err, Result};
 use diem_crypto::{hash::CryptoHash, HashValue};
-use diem_nibble::Nibble;
 use diem_types::{
+    nibble::{
+        nibble_path::{skip_common_prefix, NibbleIterator, NibblePath},
+        Nibble, ROOT_NIBBLE_HEIGHT,
+    },
     proof::{SparseMerkleProof, SparseMerkleRangeProof},
     transaction::Version,
 };
-use nibble_path::{skip_common_prefix, NibbleIterator, NibblePath};
 use node_type::{Child, Children, InternalNode, LeafNode, Node, NodeKey};
 #[cfg(any(test, feature = "fuzzing"))]
 use proptest::arbitrary::Arbitrary;
@@ -96,7 +97,7 @@ use proptest::arbitrary::Arbitrary;
 use proptest_derive::Arbitrary;
 use serde::{de::DeserializeOwned, Serialize};
 use std::{
-    collections::{BTreeMap, BTreeSet},
+    collections::{BTreeMap, BTreeSet, HashMap},
     marker::PhantomData,
 };
 use thiserror::Error;
@@ -107,9 +108,6 @@ use tree_cache::TreeCache;
 pub struct MissingRootError {
     pub version: Version,
 }
-
-/// The hardcoded maximum height of a [`JellyfishMerkleTree`] in nibbles.
-pub const ROOT_NIBBLE_HEIGHT: usize = HashValue::LENGTH * 2;
 
 /// `TreeReader` defines the interface between
 /// [`JellyfishMerkleTree`](struct.JellyfishMerkleTree.html)
@@ -185,6 +183,50 @@ pub struct TreeUpdateBatch<V> {
     pub node_stats: Vec<NodeStats>,
 }
 
+/// An iterator that iterates the index range (inclusive) of each different nibble at given
+/// `nibble_idx` of all the keys in a sorted key-value pairs.
+struct NibbleRangeIterator<'a, V> {
+    sorted_kvs: &'a [(HashValue, V)],
+    nibble_idx: usize,
+    pos: usize,
+}
+
+impl<'a, V> NibbleRangeIterator<'a, V> {
+    fn new(sorted_kvs: &'a [(HashValue, V)], nibble_idx: usize) -> Self {
+        assert!(nibble_idx < ROOT_NIBBLE_HEIGHT);
+        NibbleRangeIterator {
+            sorted_kvs,
+            nibble_idx,
+            pos: 0,
+        }
+    }
+}
+
+impl<'a, V> std::iter::Iterator for NibbleRangeIterator<'a, V> {
+    type Item = (usize, usize);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let left = self.pos;
+        if self.pos < self.sorted_kvs.len() {
+            let cur_nibble: u8 = self.sorted_kvs[left].0.nibble(self.nibble_idx);
+            let (mut i, mut j) = (left, self.sorted_kvs.len() - 1);
+            // Find the last index of the cur_nibble.
+            while i < j {
+                let mid = j - (j - i) / 2;
+                if self.sorted_kvs[mid].0.nibble(self.nibble_idx) > cur_nibble {
+                    j = mid - 1;
+                } else {
+                    i = mid;
+                }
+            }
+            self.pos = i + 1;
+            Some((left, i))
+        } else {
+            None
+        }
+    }
+}
+
 /// The Jellyfish Merkle tree data structure. See [`crate`] for description.
 pub struct JellyfishMerkleTree<'a, R, V> {
     reader: &'a R,
@@ -204,6 +246,281 @@ where
         }
     }
 
+    /// Get the node hash from the cache if exists, otherwise compute it.
+    fn get_hash(
+        node_key: &NodeKey,
+        node: &Node<V>,
+        hash_cache: &Option<&HashMap<NibblePath, HashValue>>,
+    ) -> HashValue {
+        if let Some(cache) = hash_cache {
+            match cache.get(node_key.nibble_path()) {
+                Some(hash) => *hash,
+                None => unreachable!("{:?} can not be found in hash cache", node_key),
+            }
+        } else {
+            node.hash()
+        }
+    }
+
+    /// The batch version of `put_value_sets`.
+    pub fn batch_put_value_sets(
+        &self,
+        value_sets: Vec<Vec<(HashValue, V)>>,
+        node_hashes: Option<Vec<&HashMap<NibblePath, HashValue>>>,
+        first_version: Version,
+    ) -> Result<(Vec<HashValue>, TreeUpdateBatch<V>)> {
+        let mut tree_cache = TreeCache::new(self.reader, first_version)?;
+        let hash_sets: Vec<_> = match node_hashes {
+            Some(hashes) => hashes.into_iter().map(Some).collect(),
+            None => (0..value_sets.len()).map(|_| None).collect(),
+        };
+
+        for (idx, (value_set, hash_set)) in
+            itertools::zip_eq(value_sets.into_iter(), hash_sets.into_iter()).enumerate()
+        {
+            assert!(
+                !value_set.is_empty(),
+                "Transactions that output empty write set should not be included.",
+            );
+            let version = first_version + idx as u64;
+            let deduped_and_sorted_kvs = value_set
+                .into_iter()
+                .collect::<BTreeMap<_, _>>()
+                .into_iter()
+                .collect::<Vec<_>>();
+            let root_node_key = tree_cache.get_root_node_key().clone();
+            let (new_root_node_key, _) = Self::batch_insert_at(
+                root_node_key,
+                version,
+                deduped_and_sorted_kvs.as_slice(),
+                0,
+                &hash_set,
+                &mut tree_cache,
+            )?;
+            tree_cache.set_root_node_key(new_root_node_key);
+
+            // Freezes the current cache to make all contents in the current cache immutable.
+            tree_cache.freeze();
+        }
+
+        Ok(tree_cache.into())
+    }
+
+    fn batch_insert_at(
+        mut node_key: NodeKey,
+        version: Version,
+        kvs: &[(HashValue, V)],
+        depth: usize,
+        hash_cache: &Option<&HashMap<NibblePath, HashValue>>,
+        tree_cache: &mut TreeCache<R, V>,
+    ) -> Result<(NodeKey, Node<V>)> {
+        assert!(!kvs.is_empty());
+
+        let node = tree_cache.get_node(&node_key)?;
+        Ok(match node {
+            Node::Internal(internal_node) => {
+                // We always delete the existing internal node here because it will not be referenced anyway
+                // since this version.
+                tree_cache.delete_node(&node_key, false /* is_leaf */);
+
+                // Reuse the current `InternalNode` in memory to create a new internal node.
+                let mut children: Children = internal_node.clone().into();
+
+                // Traverse all the path touched by `kvs` from this internal node.
+                for (left, right) in NibbleRangeIterator::new(kvs, depth) {
+                    // Traverse downwards from this internal node recursively by splitting the updates into
+                    // each child index
+                    let child_index = kvs[left].0.get_nibble(depth);
+
+                    let (new_child_node_key, new_child_node) =
+                        match internal_node.child(child_index) {
+                            Some(child) => {
+                                let child_node_key =
+                                    node_key.gen_child_node_key(child.version, child_index);
+                                Self::batch_insert_at(
+                                    child_node_key,
+                                    version,
+                                    &kvs[left..=right],
+                                    depth + 1,
+                                    hash_cache,
+                                    tree_cache,
+                                )?
+                            }
+                            None => {
+                                let new_child_node_key =
+                                    node_key.gen_child_node_key(version, child_index);
+                                Self::batch_create_subtree(
+                                    new_child_node_key,
+                                    version,
+                                    &kvs[left..=right],
+                                    depth + 1,
+                                    hash_cache,
+                                    tree_cache,
+                                )?
+                            }
+                        };
+
+                    children.insert(
+                        child_index,
+                        Child::new(
+                            Self::get_hash(&new_child_node_key, &new_child_node, hash_cache),
+                            version,
+                            new_child_node.is_leaf(),
+                        ),
+                    );
+                }
+                let new_internal_node = InternalNode::new(children);
+
+                node_key.set_version(version);
+
+                // Cache this new internal node.
+                tree_cache.put_node(node_key.clone(), new_internal_node.clone().into())?;
+                (node_key, new_internal_node.into())
+            }
+            Node::Leaf(leaf_node) => {
+                // We are on a leaf node but trying to insert another node, so we may diverge.
+                // We always delete the existing leaf node here because it will not be referenced anyway
+                // since this version.
+                tree_cache.delete_node(&node_key, true /* is_leaf */);
+                node_key.set_version(version);
+                Self::batch_create_subtree_with_existing_leaf(
+                    node_key, version, leaf_node, kvs, depth, hash_cache, tree_cache,
+                )?
+            }
+            Node::Null => {
+                if !node_key.nibble_path().is_empty() {
+                    bail!(
+                        "Null node exists for non-root node with node_key {:?}",
+                        node_key
+                    );
+                }
+
+                if node_key.version() == version {
+                    tree_cache.delete_node(&node_key, false /* is_leaf */);
+                }
+                Self::batch_create_subtree(
+                    NodeKey::new_empty_path(version),
+                    version,
+                    kvs,
+                    depth,
+                    hash_cache,
+                    tree_cache,
+                )?
+            }
+        })
+    }
+
+    fn batch_create_subtree_with_existing_leaf(
+        node_key: NodeKey,
+        version: Version,
+        existing_leaf_node: LeafNode<V>,
+        kvs: &[(HashValue, V)],
+        depth: usize,
+        hash_cache: &Option<&HashMap<NibblePath, HashValue>>,
+        tree_cache: &mut TreeCache<R, V>,
+    ) -> Result<(NodeKey, Node<V>)> {
+        let existing_leaf_key = existing_leaf_node.account_key();
+
+        if kvs.len() == 1 && kvs[0].0 == existing_leaf_key {
+            let new_leaf_node = Node::new_leaf(existing_leaf_key, kvs[0].1.clone());
+            tree_cache.put_node(node_key.clone(), new_leaf_node.clone())?;
+            Ok((node_key, new_leaf_node))
+        } else {
+            let existing_leaf_bucket = existing_leaf_key.get_nibble(depth);
+            let mut isolated_existing_leaf = true;
+            let mut children = Children::new();
+            for (left, right) in NibbleRangeIterator::new(kvs, depth) {
+                let child_index = kvs[left].0.get_nibble(depth);
+                let child_node_key = node_key.gen_child_node_key(version, child_index);
+                let (new_child_node_key, new_child_node) = if existing_leaf_bucket == child_index {
+                    isolated_existing_leaf = false;
+                    Self::batch_create_subtree_with_existing_leaf(
+                        child_node_key,
+                        version,
+                        existing_leaf_node.clone(),
+                        &kvs[left..=right],
+                        depth + 1,
+                        hash_cache,
+                        tree_cache,
+                    )?
+                } else {
+                    Self::batch_create_subtree(
+                        child_node_key,
+                        version,
+                        &kvs[left..=right],
+                        depth + 1,
+                        hash_cache,
+                        tree_cache,
+                    )?
+                };
+                children.insert(
+                    child_index,
+                    Child::new(
+                        Self::get_hash(&new_child_node_key, &new_child_node, hash_cache),
+                        version,
+                        new_child_node.is_leaf(),
+                    ),
+                );
+            }
+            if isolated_existing_leaf {
+                let existing_leaf_node_key =
+                    node_key.gen_child_node_key(version, existing_leaf_bucket);
+                children.insert(
+                    existing_leaf_bucket,
+                    Child::new(existing_leaf_node.hash(), version, true /* is_leaf */),
+                );
+
+                tree_cache.put_node(existing_leaf_node_key, existing_leaf_node.into())?;
+            }
+
+            let new_internal_node = InternalNode::new(children);
+
+            tree_cache.put_node(node_key.clone(), new_internal_node.clone().into())?;
+            Ok((node_key, new_internal_node.into()))
+        }
+    }
+
+    fn batch_create_subtree(
+        node_key: NodeKey,
+        version: Version,
+        kvs: &[(HashValue, V)],
+        depth: usize,
+        hash_cache: &Option<&HashMap<NibblePath, HashValue>>,
+        tree_cache: &mut TreeCache<R, V>,
+    ) -> Result<(NodeKey, Node<V>)> {
+        if kvs.len() == 1 {
+            let new_leaf_node = Node::new_leaf(kvs[0].0, kvs[0].1.clone());
+            tree_cache.put_node(node_key.clone(), new_leaf_node.clone())?;
+            Ok((node_key, new_leaf_node))
+        } else {
+            let mut children = Children::new();
+            for (left, right) in NibbleRangeIterator::new(kvs, depth) {
+                let child_index = kvs[left].0.get_nibble(depth);
+                let child_node_key = node_key.gen_child_node_key(version, child_index);
+                let (new_child_node_key, new_child_node) = Self::batch_create_subtree(
+                    child_node_key,
+                    version,
+                    &kvs[left..=right],
+                    depth + 1,
+                    hash_cache,
+                    tree_cache,
+                )?;
+                children.insert(
+                    child_index,
+                    Child::new(
+                        Self::get_hash(&new_child_node_key, &new_child_node, hash_cache),
+                        version,
+                        new_child_node.is_leaf(),
+                    ),
+                );
+            }
+            let new_internal_node = InternalNode::new(children);
+
+            tree_cache.put_node(node_key.clone(), new_internal_node.clone().into())?;
+            Ok((node_key, new_internal_node.into()))
+        }
+    }
+
     /// This is a convenient function that calls
     /// [`put_value_sets`](struct.JellyfishMerkleTree.html#method.put_value_sets) with a single
     /// `keyed_value_set`.
@@ -213,7 +530,8 @@ where
         value_set: Vec<(HashValue, V)>,
         version: Version,
     ) -> Result<(HashValue, TreeUpdateBatch<V>)> {
-        let (root_hashes, tree_update_batch) = self.put_value_sets(vec![value_set], version)?;
+        let (root_hashes, tree_update_batch) =
+            self.batch_put_value_sets(vec![value_set], None, version)?;
         assert_eq!(
             root_hashes.len(),
             1,
@@ -492,7 +810,7 @@ where
         );
 
         let internal_node = InternalNode::new(children);
-        let mut next_internal_node = internal_node.clone();
+        let mut next_internal_node: Node<V> = internal_node.clone().into();
         tree_cache.put_node(node_key.clone(), internal_node.into())?;
 
         for _i in 0..num_common_nibbles_below_internal {
@@ -506,11 +824,11 @@ where
                 Child::new(next_internal_node.hash(), version, false /* is_leaf */),
             );
             let internal_node = InternalNode::new(children);
-            next_internal_node = internal_node.clone();
+            next_internal_node = internal_node.clone().into();
             tree_cache.put_node(node_key.clone(), internal_node.into())?;
         }
 
-        Ok((node_key, next_internal_node.into()))
+        Ok((node_key, next_internal_node))
     }
 
     /// Helper function for creating leaf nodes. Returns the newly created leaf node.
@@ -675,7 +993,7 @@ impl NibbleExt for HashValue {
 mod test {
     use super::NibbleExt;
     use diem_crypto::hash::{HashValue, TestOnlyHash};
-    use diem_nibble::Nibble;
+    use diem_types::nibble::Nibble;
 
     #[test]
     fn test_common_prefix_nibbles_len() {