## Motivation

This PR adds two optimizations.
1. Jellyfish merkle tree batch update. Previously we update the tree with (address_hash, new_value) one by one. Now we merge update with same prefix following a recursive way.
2. Based on 1, now it is feasible to leverage the node digest already calculated in scratchpad, which could potentially save a lot of cpu resource, in turn, improve committing speed.

We did a simple benchmark that compared `save_transaction` api in diemDB before and after the PR.
The benchmark shows 30% - 40% speed up on average on my local machine with 
` cargo run -- --block-size=1000 --num-accounts=1000000 --num-transfer-blocks=20 ` of executor-bench
before PR: 3500-6000ms
after PR: 2000-3000ms

EXECUTOR BENCH   after->before
enchmarking bench_p2p_small: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.9s.
bench_p2p_small         time:   [522.56 ms 547.80 ms 577.81 ms]                          
                        change: [+35.918% +42.415% +49.323%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high mild

Benchmarking bench_p2p_medium: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 10.6s.
bench_p2p_medium        time:   [929.20 ms 957.67 ms 991.20 ms]                           
                        change: [+23.720% +28.121% +33.264%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 10 measurements (30.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild
  1 (10.00%) high severe

Benchmarking bench_p2p_large: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 9.5s.
bench_p2p_large         time:   [885.92 ms 905.56 ms 928.70 ms]                          
                        change: [+21.400% +24.970% +28.940%] (p = 0.00 < 0.05)
                        Performance has regressed.

### Have you read the [Contributing Guidelines on pull requests]

Y

## Test Plan

For Jellyfish merkle, update all the tests to use the new batch method and only leave one test to verify the `batch` and `serial` methods are consistent.

For leveraging exec result, the final root node will be recomputed when commit to db, which give more confidence that the code is correct when assert the accu_root from executor is equal to that calculated when committing to db.