@libra-action 2021-06-18
**release binaries dependency change summary**:
```
target packages:
  R diem-nibble 0.1.0 (workspace, path 'common/nibble')
    * (old features: default)
```
<!-- metadata: {"tag":"release-summary"} -->