@lightmark 2021-07-09
> > For leveraging exec result, the final root node will be recomputed when commit to db, which give more confidence that the code is correct when assert the accu_root from executor is equal to that calculated when committing to db.
> 
> TBH this confidence is a little bit weird IIUC. Are you saying we deliberately verify the root hash is correctly out of the up to 16 children nodes? If so I don't think it's necessary.
> 
> > The benchmark shows 30% - 40% speed up on average on my local machine.
> 
> Could you elaborate a little bit about the benchmarking setup?
> 
> And running the old executor-benchmark, did we see any gain?

Yeah, I was wrong about the confidence.