@msmouse 2021-07-09
So looks like it's 25~35% improvement of block execution+committing performance (despite of the re-traversal of the in-mem trees). Great work!